server/requirements.txt
django
djangorestframework
kisscartoon-api
gunicorn
